Ticket add

+++ lib/blocs/postBlocs/credentials/credentialBloc.dart
@@ -57,29 +57,46 @@ class CredentialsBloc extends Object with Validators {
         validName, validEmail, validPassword, validPhoneNo);
     var results = json.decode(jsonResponse);
     print(results);
-    if (results['success']) {
-      secureStorage.deleteAll();
-      secureStorage.write(key: 'id', value: results['id'].toString());
-      secureStorage.write(key: 'token', value: results['token']);
 
-      return true;
+    if (results['message'] == null) {
+      if (results['success']) {
+        secureStorage.deleteAll();
+        secureStorage.write(key: 'id', value: results['id'].toString());
+        secureStorage.write(key: 'token', value: results['token']);
+
+        return true;
+      }
+      return false;
     }
-    return false;
+    return null;
   }
 
   logout() async {
+    print('WADUUUUUU');
     String valueOfId = await secureStorage.read(key: 'id');
     String token = await secureStorage.read(key: 'token');
     int id = int.parse(valueOfId);
     var jsonResponse = await _auth.logout(id, token);
     var results = json.decode(jsonResponse);
-    print(results);
+    secureStorage.deleteAll();
+    print(results['success']);
   }
 
-  dispose() {
+  removeValues() {
+    _email.value = '';
+    _password.value = '';
+    _name.value = '';
+    _phoneNo.value = '';
+  }
+
+  disposeStreams() {
+    _email.drain();
     _email.close();
+    _password.drain();
     _password.close();
+    _name.drain();
     _name.close();
+    _phoneNo.drain();
     _phoneNo.close();
   }
 }
